Fixed window programs to implement QMainWindow
In order to add QMainWindow, I had to delete and change the order of certain things. Mainly of which was the deletion of the layout (seeing that QMainWindow would now be controlling the window). I also was able to fix the Menu and the Tool Bar and moved the canvas back to the bottom of the screen instead of the bottom 1/2.

I, however, broke the switch statement in canvas.cpp that allowed for the switching between the shapes